requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==3.0.1
scikit-learn==1.3.2
Pillow==10.1.0
python-slugify==8.0.1
